# src/main/resources/openapi-modelo_negocio.yaml
#openapi: 3.0.3
#info:
#  title: Microservicio de Clientes
#  description: API para la gestión de clientes del sistema
#  version: 1.0.0
#
#servers:
#  - url: http://localhost:8080
#
#paths:
#  /clientes:
#    get:
#      summary: Obtener lista de clientes
#      operationId: listarClientes
#      responses:
#        '200':
#          description: Lista de clientes
#          content:
#            application/json:
#              schema:
#                type: array
#                items:
#                  $ref: '#/components/schemas/Cliente'
#
#components:
#  schemas:
#    Cliente:
#      type: object
#      properties:
#        id:
#          type: integer
#          example: 1
#        nombre:
#          type: string
#          example: Juan
#        email:
#          type: string
#          example: juan@email.com
openapi: 3.0.3
info:
  title: Microservicio de Clientes
  description: API para la gestión de clientes del sistema
  version: 1.0.0

servers:
  - url: http://localhost:8080/api

paths:
  /clientes:
    get:
      summary: Obtener lista de clientes
      operationId: listarClientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
    post:
      summary: Crear un nuevo cliente
      operationId: crearCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'

  /clientes/{id}:
    get:
      summary: Obtener un cliente por ID
      operationId: obtenerCliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado
    put:
      summary: Actualizar un cliente por ID
      operationId: actualizarCliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '200':
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado
    delete:
      summary: Eliminar un cliente por ID
      operationId: eliminarCliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Cliente eliminado

components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          example: juan.perez@email.com
        telefono:
          type: string
          example: "0998765432"
        direccion:
          type: string
          example: "Av. Siempre Viva 123"
        cedula:
          type: string
          example: "1102233445"
        edad:
          type: integer
          example: 30
        ciudad:
          type: string
          example: Quito
        fechaRegistro:
          type: string
          format: date
          example: 2026-01-15

    ClienteInput:
      type: object
      required:
        - nombre
        - email
      properties:
        nombre:
          type: string
          example: Juan Pérez
        email:
          type: string
          example: juan.perez@email.com
        telefono:
          type: string
          example: "0998765432"
        direccion:
          type: string
          example: "Av. Siempre Viva 123"
        cedula:
          type: string
          example: "1102233445"
        edad:
          type: integer
          example: 30
        ciudad:
          type: string
          example: Quito